Modified benchmarking streamlit pages to support page swtiching prevention in setup. Changed app to integrate `env_input_fields` from utils.visual
