+++ benchmarking/streamlit/pages/synthetic_performance_eval_st.py
@@ -5,6 +5,7 @@
 # sys.path.append(kit_dir)
 # sys.path.append(repo_dir)
 
+import yaml
 import warnings
 
 import pandas as pd
@@ -25,11 +26,11 @@ from benchmarking.streamlit.streamlit_utils import (
 
 warnings.filterwarnings('ignore')
 
-
-@st.cache_data
-def _init() -> None:
-    load_dotenv('../.env', override=True)
-
+CONFIG_PATH = './config.yaml'
+with open(CONFIG_PATH) as file:
+    st.session_state.config = yaml.safe_load(file)
+    st.session_state.prod_mode = st.session_state.config['prod_mode']
+    st.session_state.pages_to_show = st.session_state.config['pages_to_show']
 
 def _initialize_session_variables() -> None:
     # Initialize llm
@@ -49,8 +50,6 @@ def _initialize_session_variables() -> None:
         st.session_state.timeout = None
     if 'llm_api' not in st.session_state:
         st.session_state.llm_api = None
-    if 'prod_mode' not in st.session_state:
-        st.session_state.prod_mode = None
     if 'setup_complete' not in st.session_state:
         st.session_state.setup_complete = None
 
@@ -245,7 +244,6 @@ if __name__ == '__main__':
         page_icon='https://sambanova.ai/hubfs/logotype_sambanova_orange.png',
     )
 
-    _init()
     _initialize_session_variables()
 
     if st.session_state.prod_mode:
